refactor(mobile): pre-review fixes for the ingredient readiness checklist
- Move formatQuantity to core/utils/format.ts (re-exported from recipe-details
  .utils) so the application use-case no longer imports the presentation layer
  (Clean Architecture dependency rule).
- ChecklistRow: React import first; domain ChecklistItem/ReadinessChecklist as
  interface (domain object-shape convention).
- Tests: type the readiness fixtures as RecipeDetailsIngredientItem[]; add a
  RecipeDetailsScreen test asserting the Vue-tab entry point routes to
  /recipes/[id]/readiness.
