+++ packages/mobile-app/src/core/ui/ChecklistRow.tsx
@@ -1,7 +1,7 @@
+import React from "react";
 import { Pressable, StyleSheet, Text, View } from "react-native";
 
 import { Ionicons } from "@expo/vector-icons";
-import React from "react";
 
 import { colors, spacing, typography } from "@/core/theme";
